Create CODE
The objective of Hotel Management system is to develop a project which maintains booking of rooms and handles accounts of a hotel. In this software, the hotel manager can add new customers and manage that customer information. Restaurant manager can add restaurant bill in that customer bill. This software works at all hotels that cater to a large number of customers and want to computerize the complete management process.

This software is aimed at facilitating the prompt, accurate and easy management of a hotel. It will be very helpful to the manager on the one hand and to the customers as they would be able to get good and fast service.